+++ Changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.33
+* [NEW] Warning in log and GUI if Storage Drawers version differs from the one targeted
+* [FIXED] Compatibility with Storage Drawers 3.4.4 (Belgabor/DrawersBits#6)
+
 ## Version 0.32
 * [FIXED] Compatibility with Storage Drawers 3.3.0 (Belgabor/DrawersBits#5)